Add a version of middleman-livereload for the new 3.1 extension mechanism. Incidentally fixes #15
