+++ lib/middleman-livereload/version.rb
@@ -1,5 +1,5 @@
 module Middleman
   module LiveReload
-    VERSION = "3.0.1"
+    VERSION = "3.1.0"
   end
 end
