+++ middleman-livereload.gemspec
@@ -18,7 +18,7 @@ Gem::Specification.new do |s|
   s.test_files    = `git ls-files -z -- {fixtures,features}/*`.split("\0")
   s.require_paths = ["lib"]
   
-  s.add_dependency("middleman-core", ["~> 3.0.2"])
+  s.add_dependency("middleman-core", ["~> 3.0"])
   s.add_runtime_dependency('rack-livereload')
   s.add_runtime_dependency('em-websocket', ['>= 0.2.0'])
   s.add_runtime_dependency('multi_json', ['~> 1.0'])
